# Optimise suffix_index() for URLs with long subdomains