--- a/tldextract/tldextract.py
+++ b/tldextract/tldextract.py
@@ -88,8 +88,8 @@ class ExtractResult(NamedTuple):
         >>> extract('http://localhost:8080').registered_domain
         ''
         """
-        if self.suffix and self.domain:
-            return f"{self.domain}.{self.suffix}"
+        if self.domain and self.suffix:
+            return self.domain + "." + self.suffix
         return ""
 
     @property
@@ -102,7 +102,7 @@ class ExtractResult(NamedTuple):
         >>> extract('http://localhost:8080').fqdn
         ''
         """
-        if self.suffix and self.domain:
+        if self.domain and self.suffix:
             # Disable bogus lint error (https://github.com/PyCQA/pylint/issues/2568)
             # pylint: disable-next=not-an-iterable
             return ".".join(i for i in self if i)